+++ src/main/java/dev/the_fireplace/caterpillar/client/screen/widget/TutorialButton.java
@@ -42,11 +42,11 @@ public class TutorialButton extends ImageButton {
         super.blit(poseStack, super.getX(), super.getY(), this.xTexStart, yOffset, super.width, super.height);
     }
 
-    public void setShowTutorial(boolean showTutorial) {
-        this.showTutorial = showTutorial;
+    public void toggleTutorial() {
+        this.showTutorial = !this.showTutorial;
     }
 
-    public boolean showTutorial() {
+    public boolean isTutorialShown() {
         return this.showTutorial;
     }
 }